+++ nellai_iptv_app/lib/models/public_settings.dart
@@ -1,13 +1,17 @@
+import 'package:flutter/foundation.dart';
+
 class PublicSettings {
   final String? appName;
   final String? logoUrl;
   final String? appLogoPngUrl;
+  final String? fallbackHlsUrl;
   final String? favicon;
 
   PublicSettings({
     this.appName,
     this.logoUrl,
     this.appLogoPngUrl,
+    this.fallbackHlsUrl,
     this.favicon,
   });
 
@@ -15,7 +19,8 @@ class PublicSettings {
     return PublicSettings(
       appName: json['app_name'],
       logoUrl: json['logo_url'] ?? json['logo'],
-      appLogoPngUrl: json['app_logo_png_url'], // New field
+      appLogoPngUrl: json['app_logo_png_url'],
+      fallbackHlsUrl: json['fallback_404_hls_url'],
       favicon: json['favicon'],
     );
   }